# [Linux-SGX PAL] Add exitless system calls