--- a/Pal/src/host/Linux-SGX/sgx_rpcqueue.c
+++ b/Pal/src/host/Linux-SGX/sgx_rpcqueue.c
@@ -1,0 +1,4 @@
+#include "rpcqueue.h"
+
+/* duplicates functionality outside of enclave */
+#include "enclave_rpcqueue.c"