… Pal/src/host/Linux-SGX/enclave_ocalls.c