This PR succeeds PR #235.

This PR adds the ability to invoke system calls in an "exitless" fashion (aka asynchronously), similar to solutions from [Eleos by Orenbach](https://dl.acm.org/citation.cfm?id=3064219) and [SCONE by Arnautov et al](https://dl.acm.org/citation.cfm?id=3026930).

New directive `sgx.rpc_thread_num = X` in the manifest instructs Graphene-SGX to create X outside-enclave RPC threads that spin-wait for syscalls (and any OCALLs in general) on a shared queue. In-enclave threads do not exit the enclave now but instead enqueue the syscall in the queue and wait for it to be processed by an RPC thread. Not to waste CPU resources, enclave threads first spin (in the hope syscall will return immediately) and then sleep on a futex.

This PR reserves the SIGUSR2 signal for internal usage. In particular, when a signal arrives to the enclave application (e.g. SIGINT or SIGCONT), RPC threads are notified using SIGUSR2. This notification is required in case RPC threads are blocked on system calls; POSIX mandates that signals interrupt such blocked system calls ([`signal` manpage](http://man7.org/linux/man-pages/man7/signal.7.html)).

Omitting `sgx.rpc_thread_num` defaults to the old way of executing syscalls: exiting the enclave, executing the syscall, and re-entering it again.

This PR is stable. It works with multi-process applications and correctly handles signals.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/oscarlab/graphene/405)
<!-- Reviewable:end -->