@dimakuv 2018-12-26
Some performance numbers to showcase better performance:

```
Original Graphene
-----------------

LMBENCH:
NULL:  Simple syscall median=0.0678 [mean=0.0678 +/-0.0006] microseconds
OPEN:  Simple open/close median=33.9811 [mean=34.1377 +/-0.2349] microseconds
READ:  Simple read median=0.3314 [mean=0.3321 +/-0.0011] microseconds
WRITE: Simple write median=0.3205 [mean=0.3207 +/-0.0007] microseconds
STAT:  Simple stat median=2.2327 [mean=2.3136 +/-0.1724] microseconds


MEMCACHED 1.4.33:
(server: SGX=1 ./pal_loader ./memcached -u <myuser> -t 1)
(client: memtier_benchmark -P memcache_text -p 11211 --test-time 30)

Type        Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
------------------------------------------------------------------------
Sets         676.62          ---          ---    292.66300        46.18
Gets        6699.82         6.76      6693.07      0.27400       169.79
Waits          0.00          ---          ---      0.00000          ---
Totals      7376.44         6.76      6693.07     27.09300       215.97


REDIS 5.0:
(server: SGX=1 ./pal_loader ./redis-server --save '' --protected-mode no)
(client: memtier_benchmark --test-time 30)

Type        Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
------------------------------------------------------------------------
Sets         563.41          ---          ---    345.42800        43.40
Gets        5568.01         3.73      5564.28      0.96800       189.85
Waits          0.00          ---          ---      0.00000          ---
Totals      6131.42         3.73      5564.28     32.62100       233.25



Exitless Graphene
-----------------

LMBENCH:
NULL:  Simple syscall median=0.0724 [mean=0.0726 +/-0.0002] microseconds
OPEN:  Simple open/close median=11.3532 [mean=11.3600 +/-0.0924] microseconds
READ:  Simple read median=0.3554 [mean=0.3549 +/-0.0009] microseconds
WRITE: Simple write median=0.3452 [mean=0.3453 +/-0.0008] microseconds
STAT:  Simple stat median=2.3863 [mean=2.3874 +/-0.0051] microseconds


MEMCACHED 1.4.33:
(server: SGX=1 ./pal_loader ./memcached -u <myuser> -t 1)
(client: memtier_benchmark -P memcache_text -p 11211 --test-time 30)

Type        Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
------------------------------------------------------------------------
Sets         861.43          ---          ---    183.72700        58.79
Gets        8556.62        10.79      8545.83      4.86300       216.98
Waits          0.00          ---          ---      0.00000          ---
Totals      9418.04        10.79      8545.83     21.22300       275.76


REDIS 5.0:
(server: SGX=1 ./pal_loader ./redis-server --save '' --protected-mode no)
(client: memtier_benchmark --test-time 30)

Type        Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
------------------------------------------------------------------------
Sets        1087.80          ---          ---    183.54000        83.80
Gets       10811.07        11.58     10799.49      0.66600       368.78
Waits          0.00          ---          ---      0.00000          ---
Totals     11898.87        11.58     10799.49     17.38500       452.58
```

Summary:

* With lmbench, `open` syscall is *3x faster*;

* With Memcached, total throughput is *28%* higher;

* With Redis, total throughput is *94%* higher.